[Decomp][Spell] Cohesion split: extract damage/trigger/teleport handlers to SpellEffectDamageTeleport.cpp
Moves the head-of-file cluster of 12 direct-damage, trigger-spell and
jump/teleport effect handlers (EffectResurrectNew, EffectInstaKill,
EffectEnvironmentalDMG, EffectSchoolDMG, EffectTriggerSpellWithValue,
EffectTriggerRitualOfSummoning, EffectClearQuest, EffectForceCast,
EffectTriggerSpell, EffectTriggerMissileSpell, EffectJump, EffectTeleportUnits)
into a dedicated SpellEffectDamageTeleport.cpp. Declarations stay in Spell.h;
the SpellEffects[] dispatch table is untouched and resolves the handlers by
address at link time. SpellEffects.cpp 10,708 -> 9,508. No behaviour change.
