+++ src/game/WorldHandlers/SpellEffects.cpp
@@ -290,1206 +290,6 @@ void Spell::EffectUnused(SpellEffectEntry const* /*effect*/)
 }
 
 /**
- * @brief Sends a resurrection request to a dead player target.
- *
- * @param effect The effect index providing resurrection values.
- */
-void Spell::EffectResurrectNew(SpellEffectEntry const* effect)
-{
-    if (!unitTarget || unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    if (!unitTarget->IsInWorld())
-    {
-        return;
-    }
-
-    Player* pTarget = ((Player*)unitTarget);
-
-    if (pTarget->isRessurectRequested())      // already have one active request
-    {
-        return;
-    }
-
-    uint32 health = damage;
-    uint32 mana = effect->EffectMiscValue;
-    pTarget->setResurrectRequestData(m_caster->GetObjectGuid(), m_caster->GetMapId(), m_caster->GetPositionX(), m_caster->GetPositionY(), m_caster->GetPositionZ(), health, mana);
-    SendResurrectRequest(pTarget);
-}
-
-/**
- * @brief Instantly kills the unit target and handles spell-specific side effects.
- *
- * @param effect Unused effect index.
- */
-void Spell::EffectInstaKill(SpellEffectEntry const* /*effect*/)
-{
-    if (!unitTarget || !unitTarget->IsAlive())
-    {
-        return;
-    }
-
-    if (m_caster == unitTarget)                             // prevent interrupt message
-    {
-        finish();
-    }
-
-    WorldObject* caster = GetCastingObject();               // we need the original casting object
-
-    WorldPacket data(SMSG_SPELLINSTAKILLLOG, (8 + 8 + 4));
-    data << (caster && caster->GetTypeId() != TYPEID_GAMEOBJECT ? m_caster->GetObjectGuid() : ObjectGuid()); // Caster GUID
-    data << unitTarget->GetObjectGuid();                    // Victim GUID
-    data << uint32(m_spellInfo->Id);
-    m_caster->SendMessageToSet(&data, true);
-
-    m_caster->DealDamage(unitTarget, unitTarget->GetHealth(), NULL, DIRECT_DAMAGE, SPELL_SCHOOL_MASK_NORMAL, m_spellInfo, false);
-}
-
-/**
- * @brief Applies environmental damage to the caster.
- *
- * @param effect The effect index used to calculate the base damage.
- */
-void Spell::EffectEnvironmentalDMG(SpellEffectEntry const* effect)
-{
-    uint32 absorb = 0;
-    uint32 resist = 0;
-
-    // Note: this hack with damage replace required until GO casting not implemented
-    // environment damage spells already have around enemies targeting but this not help in case nonexistent GO casting support
-    // currently each enemy selected explicitly and self cast damage, we prevent apply self casted spell bonuses/etc
-    damage = effect->CalculateSimpleValue();
-
-    m_caster->CalculateDamageAbsorbAndResist(m_caster, GetSpellSchoolMask(m_spellInfo), SPELL_DIRECT_DAMAGE, damage, &absorb, &resist);
-
-    m_caster->SendSpellNonMeleeDamageLog(m_caster, m_spellInfo->Id, damage, GetSpellSchoolMask(m_spellInfo), absorb, resist, false, 0, false);
-    if (m_caster->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)m_caster)->EnvironmentalDamage(DAMAGE_FIRE, damage);
-    }
-}
-
-/**
- * @brief Computes school-damage special cases and accumulates resulting damage.
- *
- * @param effect The damage effect index.
- */
-void Spell::EffectSchoolDMG(SpellEffectEntry const* effect)
-{
-    if (unitTarget && unitTarget->IsAlive())
-    {
-        SpellClassOptionsEntry const* classOptions = m_spellInfo->GetSpellClassOptions();
-
-        switch(m_spellInfo->GetSpellFamilyName())
-        {
-            case SPELLFAMILY_GENERIC:
-            {
-                switch (m_spellInfo->Id)                    // better way to check unknown
-                {
-                    // Meteor like spells (divided damage to targets)
-                    case 24340: case 26558: case 28884:     // Meteor
-                    case 36837: case 38903: case 41276:     // Meteor
-                    case 57467:                             // Meteor
-                    case 26789:                             // Shard of the Fallen Star
-                    case 31436:                             // Malevolent Cleave
-                    case 35181:                             // Dive Bomb
-                    case 40810: case 43267: case 43268:     // Saber Lash
-                    case 42384:                             // Brutal Swipe
-                    case 45150:                             // Meteor Slash
-                    case 64422: case 64688:                 // Sonic Screech
-                    case 70492: case 72505:                 // Ooze Eruption
-                    case 71904:                             // Chaos Bane
-                    case 72624: case 72625:                 // Ooze Eruption
-                    case 77679: case 92968: case 92969: case 92970:     // Scorching Blast
-                    case 82935: case 88915: case 88916: case 88917:     // Caustic Slime
-                    case 86014: case 92863: case 92864: case 92865:     // Twilight Meteorite
-                    case 86367: case 93135: case 93136: case 93137:     // Sleet Storm
-                    case 86825: case 92879: case 92880: case 92881:     // Blackout
-                    case 88942: case 95172:                 // Meteor Slash
-                    case 89348: case 95178:                 // Demon Repellent Ray
-                    case 98474: case 100212: case 100213: case 100214:  //Flame Scythe
-                    case 103414: case 108571: case 109033: case 109034: //Stomp
-                    case 105069: case 108094:               // Seething Hate
-                    case 106375: case 109182: case 109183: case 109184: //Twilight Unstable
-                    {
-                        uint32 count = 0;
-                        for(TargetList::const_iterator ihit = m_UniqueTargetInfo.begin(); ihit != m_UniqueTargetInfo.end(); ++ihit)
-                            if (ihit->effectMask & (1<<effect->EffectIndex))
-                            {
-                                ++count;
-                            }
-
-                        damage /= count;                    // divide to all targets
-                        break;
-                    }
-                    // percent from health with min
-                    case 25599:                             // Thundercrash
-                    {
-                        damage = unitTarget->GetHealth() / 2;
-                        if (damage < 200)
-                        {
-                            damage = 200;
-                        }
-                        break;
-                    }
-                    // Intercept (warrior spell trigger)
-                    case 20253:
-                    case 61491:
-                    {
-                        damage += uint32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.12f);
-                        break;
-                    }
-                    // percent max target health
-                    case 29142:                             // Eyesore Blaster
-                    case 35139:                             // Throw Boom's Doom
-                    case 49882:                             // Leviroth Self-Impale
-                    case 55269:                             // Deathly Stare
-                    {
-                        damage = damage * unitTarget->GetMaxHealth() / 100;
-                        break;
-                    }
-                    // Lightning Strike
-                    case 37841:
-                        if (unitTarget->GetTypeId() == TYPEID_PLAYER && unitTarget->HasAura(37830)) // Repolarized Magneto Sphere
-                        {
-                            ((Player*)unitTarget)->KilledMonsterCredit(21910);
-                        }
-                        break;
-                    // Cataclysmic Bolt
-                    // Cataclysmic Bolt
-                    case 38441:
-                    {
-                        damage = unitTarget->GetMaxHealth() / 2;
-                        break;
-                    }
-                    // Touch the Nightmare
-                    case 50341:
-                    {
-                        if (SpellEffectIndex(effect->EffectIndex) == EFFECT_INDEX_2)
-                        {
-                            damage = int32(unitTarget->GetMaxHealth() * 0.3f);
-                        }
-                        break;
-                    }
-                    // Tympanic Tantrum
-                    case 62775:
-                    {
-                        damage = unitTarget->GetMaxHealth() / 10;
-                        break;
-                    }
-                    // Hand of Rekoning (name not have typos ;) )
-                    case 67485:
-                        damage += uint32(0.5f * m_caster->GetTotalAttackPowerValue(BASE_ATTACK));
-                        break;
-                    // Magic Bane normal (Forge of Souls - Bronjahm)
-                    case 68793:
-                    {
-                        damage += uint32(unitTarget->GetMaxPower(POWER_MANA) / 2);
-                        damage = std::min(damage, 10000);
-                        break;
-                    }
-                    // Magic Bane heroic (Forge of Souls - Bronjahm)
-                    case 69050:
-                    {
-                        damage += uint32(unitTarget->GetMaxPower(POWER_MANA) / 2);
-                        damage = std::min(damage, 15000);
-                        break;
-                    }
-                }
-                break;
-            }
-            case SPELLFAMILY_MAGE:
-                // remove Arcane Blast buffs at any non-Arcane Blast arcane damage spell.
-                // NOTE: it removed at hit instead cast because currently spell done-damage calculated at hit instead cast
-                if ((m_spellInfo->SchoolMask & SPELL_SCHOOL_MASK_ARCANE) && !(classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x20000000)))
-                {
-                    m_caster->RemoveAurasDueToSpell(36032); // Arcane Blast buff
-                }
-                break;
-            case SPELLFAMILY_WARRIOR:
-            {
-                // Bloodthirst
-                if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x40000000000))
-                {
-                    damage = uint32(damage * (m_caster->GetTotalAttackPowerValue(BASE_ATTACK)) / 100);
-                }
-                // Victory Rush
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x10000000000))
-                {
-                    damage = uint32(damage * m_caster->GetTotalAttackPowerValue(BASE_ATTACK) / 100);
-                    m_caster->ModifyAuraState(AURA_STATE_WARRIOR_VICTORY_RUSH, false);
-                }
-                // Revenge ${$m1+$AP*0.310} to ${$M1+$AP*0.310}
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000000000400))
-                {
-                    damage+= uint32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.310f);
-                }
-                // Heroic Throw ${$m1+$AP*.50}
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000100000000))
-                {
-                    damage+= uint32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.5f);
-                }
-                // Shattering Throw ${$m1+$AP*.50}
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0040000000000000))
-                {
-                    damage+= uint32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.5f);
-                }
-                // Shockwave ${$m3/100*$AP}
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000800000000000))
-                {
-                    int32 pct = m_caster->CalculateSpellDamage(unitTarget, m_spellInfo, EFFECT_INDEX_2);
-                    if (pct > 0)
-                    {
-                        damage += int32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * pct / 100);
-                    }
-                    break;
-                }
-                // Thunder Clap
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000000000080))
-                {
-                    damage += int32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 12 / 100);
-                }
-                break;
-            }
-            case SPELLFAMILY_WARLOCK:
-            {
-                // Incinerate Rank 1 & 2
-                if ((classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x00004000000000)) && m_spellInfo->SpellIconID==2128)
-                {
-                    // Incinerate does more dmg (dmg*0.25) if the target have Immolate debuff.
-                    // Check aura state for speed but aura state set not only for Immolate spell
-                    if (unitTarget->HasAuraState(AURA_STATE_CONFLAGRATE))
-                    {
-                        Unit::AuraList const& RejorRegr = unitTarget->GetAurasByType(SPELL_AURA_PERIODIC_DAMAGE);
-                        for (Unit::AuraList::const_iterator i = RejorRegr.begin(); i != RejorRegr.end(); ++i)
-                        {
-                            // Immolate
-                            SpellClassOptionsEntry const* immSpellClassOpt = (*i)->GetSpellProto()->GetSpellClassOptions();
-                            if (!immSpellClassOpt)
-                            {
-                                continue;
-                            }
-                            if (immSpellClassOpt->SpellFamilyName == SPELLFAMILY_WARLOCK &&
-                                (immSpellClassOpt->SpellFamilyFlags & UI64LIT(0x00000000000004)))
-                            {
-                                damage += damage / 4;
-                                break;
-                            }
-                        }
-                    }
-                }
-                // Shadowflame
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0001000000000000))
-                {
-                    // Apply DOT part
-                    switch (m_spellInfo->Id)
-                    {
-                        case 47897: m_caster->CastSpell(unitTarget, 47960, true); break;
-                        case 61290: m_caster->CastSpell(unitTarget, 61291, true); break;
-                        default:
-                            sLog.outError("Spell::EffectDummy: Unhandeled Shadowflame spell rank %u", m_spellInfo->Id);
-                            break;
-                    }
-                }
-                // Shadow Bite
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0040000000000000))
-                {
-                    Unit* owner = m_caster->GetOwner();
-                    if (!owner)
-                    {
-                        break;
-                    }
-
-                    uint32 counter = 0;
-                    Unit::AuraList const& dotAuras = unitTarget->GetAurasByType(SPELL_AURA_PERIODIC_DAMAGE);
-                    for (Unit::AuraList::const_iterator itr = dotAuras.begin(); itr != dotAuras.end(); ++itr)
-                        if ((*itr)->GetCasterGuid() == owner->GetObjectGuid())
-                        {
-                            ++counter;
-                        }
-
-                    if (counter)
-                    {
-                        damage += (counter * owner->CalculateSpellDamage(unitTarget, m_spellInfo, EFFECT_INDEX_2) * damage) / 100.0f;
-                    }
-                }
-                // Conflagrate - consumes Immolate or Shadowflame
-                else if (m_spellInfo->GetTargetAuraState() == AURA_STATE_CONFLAGRATE)
-                {
-                    Aura const* aura = NULL;                // found req. aura for damage calculation
-
-                    Unit::AuraList const& mPeriodic = unitTarget->GetAurasByType(SPELL_AURA_PERIODIC_DAMAGE);
-                    for (Unit::AuraList::const_iterator i = mPeriodic.begin(); i != mPeriodic.end(); ++i)
-                    {
-                        // for caster applied auras only
-                        if ((*i)->GetSpellProto()->GetSpellFamilyName() != SPELLFAMILY_WARLOCK ||
-                            (*i)->GetCasterGuid() != m_caster->GetObjectGuid())
-                            continue;
-
-                        // Immolate
-                        if ((*i)->GetSpellProto()->IsFitToFamilyMask(UI64LIT(0x0000000000000004)))
-                        {
-                            aura = *i;                      // it selected always if exist
-                            break;
-                        }
-
-                        // Shadowflame
-                        if ((*i)->GetSpellProto()->IsFitToFamilyMask(UI64LIT(0x0000000000000000), 0x00000002))
-                        {
-                            aura = *i;                      // remember but wait possible Immolate as primary priority
-                        }
-                    }
-
-                    // found Immolate or Shadowflame
-                    if (aura)
-                    {
-                        int32 damagetick = aura->GetModifier()->m_amount;
-                        damage += damagetick * 4;
-
-                        // Glyph of Conflagrate
-                        if (!m_caster->HasAura(56235))
-                        {
-                            unitTarget->RemoveAurasByCasterSpell(aura->GetId(), m_caster->GetObjectGuid());
-                        }
-                        break;
-                    }
-                }
-                break;
-            }
-            case SPELLFAMILY_PRIEST:
-            {
-                // Shadow Word: Death - deals damage equal to damage done to caster
-                if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000200000000))
-                {
-                    m_caster->CastCustomSpell(m_caster, 32409, &damage, 0, 0, true);
-                }
-                // Improved Mind Blast (Mind Blast in shadow form bonus)
-                else if (m_caster->GetShapeshiftForm() == FORM_SHADOW && (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x00002000)))
-                {
-                    Unit::AuraList const& ImprMindBlast = m_caster->GetAurasByType(SPELL_AURA_ADD_FLAT_MODIFIER);
-                    for (Unit::AuraList::const_iterator i = ImprMindBlast.begin(); i != ImprMindBlast.end(); ++i)
-                    {
-                        if ((*i)->GetSpellProto()->GetSpellFamilyName() == SPELLFAMILY_PRIEST &&
-                            ((*i)->GetSpellProto()->SpellIconID == 95))
-                        {
-                            int chance = (*i)->GetSpellProto()->CalculateSimpleValue(EFFECT_INDEX_1);
-                            if (roll_chance_i(chance))
-                                // Mind Trauma
-                                m_caster->CastSpell(unitTarget, 48301, true);
-                            break;
-                        }
-                    }
-                }
-                break;
-            }
-            case SPELLFAMILY_DRUID:
-            {
-                SpellEffectEntry const* rakeSpellEffect = m_spellInfo->GetSpellEffect(EFFECT_INDEX_2);
-                // Ferocious Bite
-                if (m_caster->GetTypeId()==TYPEID_PLAYER && (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x000800000)) && m_spellInfo->SpellVisual[0]==6587)
-                {
-                    // converts up to 30 points of energy into ($f1+$AP/410) additional damage
-                    float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    float multiple = ap / 410 + effect->EffectDamageMultiplier;
-                    damage += int32(((Player*)m_caster)->GetComboPoints() * ap * 7 / 100);
-                    uint32 energy = m_caster->GetPower(POWER_ENERGY);
-                    uint32 used_energy = energy > 30 ? 30 : energy;
-                    damage += int32(used_energy * multiple);
-                    m_caster->SetPower(POWER_ENERGY, energy - used_energy);
-                }
-                // Rake
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000000001000) && rakeSpellEffect && rakeSpellEffect->Effect == SPELL_EFFECT_ADD_COMBO_POINTS)
-                {
-                    // $AP*0.01 bonus
-                    damage += int32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) / 100);
-                }
-                // Swipe
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0010000000000000))
-                {
-                    damage += int32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.08f);
-                }
-                break;
-            }
-            case SPELLFAMILY_ROGUE:
-            {
-                // Envenom
-                if (m_caster->GetTypeId()==TYPEID_PLAYER && (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x800000000)))
-                {
-                    // consume from stack dozes not more that have combo-points
-                    if (uint32 combo = ((Player*)m_caster)->GetComboPoints())
-                    {
-                        Aura* poison = 0;
-                        // Lookup for Deadly poison (only attacker applied)
-                        Unit::AuraList const& auras = unitTarget->GetAurasByType(SPELL_AURA_PERIODIC_DAMAGE);
-                        for(Unit::AuraList::const_iterator itr = auras.begin(); itr!=auras.end(); ++itr)
-                        {
-                            SpellClassOptionsEntry const* poisonClassOptions = (*itr)->GetSpellProto()->GetSpellClassOptions();
-                            if (!poisonClassOptions)
-                            {
-                                continue;
-                            }
-                            if ( poisonClassOptions->SpellFamilyName==SPELLFAMILY_ROGUE &&
-                                (poisonClassOptions->SpellFamilyFlags & UI64LIT(0x10000)) &&
-                                (*itr)->GetCasterGuid() == m_caster->GetObjectGuid())
-                            {
-                                poison = *itr;
-                                break;
-                            }
-                        }
-                        // count consumed deadly poison doses at target
-                        if (poison)
-                        {
-                            bool needConsume = true;
-                            uint32 spellId = poison->GetId();
-                            uint32 doses = poison->GetStackAmount();
-                            if (doses > combo)
-                            {
-                                doses = combo;
-                            }
-
-                            // Master Poisoner
-                            Unit::AuraList const& auraList = ((Player*)m_caster)->GetAurasByType(SPELL_AURA_MOD_DURATION_OF_EFFECTS_BY_DISPEL);
-                            for (Unit::AuraList::const_iterator iter = auraList.begin(); iter != auraList.end(); ++iter)
-                            {
-                                if ((*iter)->GetSpellProto()->GetSpellFamilyName() == SPELLFAMILY_ROGUE && (*iter)->GetSpellProto()->SpellIconID == 1960)
-                                {
-                                    if (int32 chance = (*iter)->GetSpellProto()->CalculateSimpleValue(EFFECT_INDEX_2))
-                                        if (roll_chance_i(chance))
-                                        {
-                                            needConsume = false;
-                                        }
-
-                                    break;
-                                }
-                            }
-
-                            if (needConsume)
-                            {
-                                unitTarget->RemoveAuraHolderFromStack(spellId, doses, m_caster->GetObjectGuid());
-                            }
-
-                            damage *= doses;
-                            damage += int32(((Player*)m_caster)->GetTotalAttackPowerValue(BASE_ATTACK) * 0.09f * doses);
-                        }
-                        // Eviscerate and Envenom Bonus Damage (item set effect)
-                        if (m_caster->GetDummyAura(37169))
-                        {
-                            damage += ((Player*)m_caster)->GetComboPoints() * 40;
-                        }
-                    }
-                }
-                // Eviscerate
-                else if ((classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x00020000)) && m_caster->GetTypeId()==TYPEID_PLAYER)
-                {
-                    if (uint32 combo = ((Player*)m_caster)->GetComboPoints())
-                    {
-                        float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                        damage += irand(int32(ap * combo * 0.03f), int32(ap * combo * 0.07f));
-
-                        // Eviscerate and Envenom Bonus Damage (item set effect)
-                        if (m_caster->GetDummyAura(37169))
-                        {
-                            damage += combo * 40;
-                        }
-                    }
-                }
-                break;
-            }
-            case SPELLFAMILY_HUNTER:
-            {
-                // Gore
-                if (m_spellInfo->SpellIconID == 1578)
-                {
-                    if (m_caster->HasAura(57627))           // Charge 6 sec post-affect
-                    {
-                        damage *= 2;
-                    }
-                }
-                // Steady Shot
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x100000000))
-                {
-                    int32 base = irand((int32)m_caster->GetWeaponDamageRange(RANGED_ATTACK, MINDAMAGE), (int32)m_caster->GetWeaponDamageRange(RANGED_ATTACK, MAXDAMAGE));
-                    uint32 attTime = m_caster->GetAttackTime(RANGED_ATTACK);
-                    damage += attTime ? int32(float(base) / attTime * 2800 + m_caster->GetTotalAttackPowerValue(RANGED_ATTACK) * 0.1f) : 0;
-                }
-                break;
-            }
-            case SPELLFAMILY_PALADIN:
-            {
-                // Judgement of Righteousness - receive benefit from Spell Damage and Attack power
-                if (m_spellInfo->Id == 20187)
-                {
-                    float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    int32 holy = m_caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(m_spellInfo));
-                    if (holy < 0)
-                    {
-                        holy = 0;
-                    }
-                    damage += int32(ap * 0.2f) + int32(holy * 32 / 100);
-                }
-                // Judgement of Vengeance/Corruption ${1+0.22*$SPH+0.14*$AP} + 10% for each application of Holy Vengeance/Blood Corruption on the target
-                else if ((classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x800000000)) && m_spellInfo->SpellIconID==2292)
-                {
-                    uint32 debuf_id;
-                    switch (m_spellInfo->Id)
-                    {
-                        case 53733: debuf_id = 53742; break;// Judgement of Corruption -> Blood Corruption
-                        case 31804: debuf_id = 31803; break;// Judgement of Vengeance -> Holy Vengeance
-                        default: return;
-                    }
-
-                    float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    int32 holy = m_caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(m_spellInfo));
-                    if (holy < 0)
-                    {
-                        holy = 0;
-                    }
-                    damage += int32(ap * 0.14f) + int32(holy * 22 / 100);
-                    // Get stack of Holy Vengeance on the target added by caster
-                    uint32 stacks = 0;
-                    Unit::AuraList const& auras = unitTarget->GetAurasByType(SPELL_AURA_PERIODIC_DAMAGE);
-                    for (Unit::AuraList::const_iterator itr = auras.begin(); itr != auras.end(); ++itr)
-                    {
-                        if (((*itr)->GetId() == debuf_id) && (*itr)->GetCasterGuid() == m_caster->GetObjectGuid())
-                        {
-                            stacks = (*itr)->GetStackAmount();
-                            break;
-                        }
-                    }
-                    // + 10% for each application of Holy Vengeance on the target
-                    if (stacks)
-                    {
-                        damage += damage * stacks * 10 / 100;
-                    }
-                }
-                // Avenger's Shield ($m1+0.07*$SPH+0.07*$AP) - ranged sdb for future
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000000000004000))
-                {
-                    float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    int32 holy = m_caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(m_spellInfo));
-                    if (holy < 0)
-                    {
-                        holy = 0;
-                    }
-                    damage += int32(ap * 0.07f) + int32(holy * 7 / 100);
-                }
-                // Hammer of Wrath ($m1+0.15*$SPH+0.15*$AP) - ranged type sdb future fix
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0000008000000000))
-                {
-                    float ap = m_caster->GetTotalAttackPowerValue(BASE_ATTACK);
-                    int32 holy = m_caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(m_spellInfo));
-                    if (holy < 0)
-                    {
-                        holy = 0;
-                    }
-                    damage += int32(ap * 0.15f) + int32(holy * 15 / 100);
-                }
-                // Hammer of the Righteous
-                else if (classOptions && classOptions->SpellFamilyFlags & UI64LIT(0x0004000000000000))
-                {
-                    // Add main hand dps * effect[2] amount
-                    float average = (m_caster->GetFloatValue(UNIT_FIELD_MINDAMAGE) + m_caster->GetFloatValue(UNIT_FIELD_MAXDAMAGE)) / 2;
-                    int32 count = m_caster->CalculateSpellDamage(unitTarget, m_spellInfo, EFFECT_INDEX_2);
-                    uint32 attTime = m_caster->GetAttackTime(BASE_ATTACK);
-                    damage += attTime ? count * int32(average * IN_MILLISECONDS) / attTime : 0;
-                }
-                // Judgement
-                else if (m_spellInfo->Id == 20271)
-                {
-                    // [1 + 0.25 * SPH + 0.16 * AP]
-                    damage += int32(m_caster->GetTotalAttackPowerValue(BASE_ATTACK) * 0.16f);
-                }
-                break;
-            }
-        }
-
-        if (damage >= 0)
-        {
-            m_damage += damage;
-        }
-    }
-}
-
-void Spell::EffectTriggerSpellWithValue(SpellEffectEntry const* effect)
-{
-    uint32 triggered_spell_id = effect->EffectTriggerSpell;
-
-    // normal case
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(triggered_spell_id);
-
-    if (!spellInfo)
-    {
-        sLog.outError("EffectTriggerSpellWithValue of spell %u: triggering unknown spell id %i", m_spellInfo->Id, triggered_spell_id);
-        return;
-    }
-
-    int32 bp = damage;
-    m_caster->CastCustomSpell(unitTarget, triggered_spell_id, &bp, &bp, &bp, true, m_CastItem , NULL, m_originalCasterGUID, m_spellInfo);
-}
-
-void Spell::EffectTriggerRitualOfSummoning(SpellEffectEntry const* effect)
-{
-    uint32 triggered_spell_id = effect->EffectTriggerSpell;
-    SpellEntry const *spellInfo = sSpellStore.LookupEntry( triggered_spell_id );
-
-    if (!spellInfo)
-    {
-        sLog.outError("EffectTriggerRitualOfSummoning of spell %u: triggering unknown spell id %i", m_spellInfo->Id, triggered_spell_id);
-        return;
-    }
-
-    finish();
-
-    m_caster->CastSpell(unitTarget, spellInfo, false);
-}
-
-void Spell::EffectClearQuest(SpellEffectEntry const* effect)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)m_caster;
-
-    uint32 quest_id = effect->EffectMiscValue;
-
-    if (!sObjectMgr.GetQuestTemplate(quest_id))
-    {
-        sLog.outError("Spell::EffectClearQuest spell entry %u attempt clear quest entry %u but this quest does not exist.", m_spellInfo->Id, quest_id);
-        return;
-    }
-
-    // remove quest possibly in quest log (is that expected?)
-    for (uint16 slot = 0; slot < MAX_QUEST_LOG_SIZE; ++slot)
-    {
-        uint32 quest = player->GetQuestSlotQuestId(slot);
-
-        if (quest == quest_id)
-        {
-            player->SetQuestSlot(slot, 0);
-            // ignore unequippable quest items in this case, it will still be equipped
-            player->TakeQuestSourceItem(quest_id, false);
-        }
-    }
-
-    player->SetQuestStatus(quest_id, QUEST_STATUS_NONE);
-    player->getQuestStatusMap()[quest_id].m_rewarded = false;
-}
-
-void Spell::EffectForceCast(SpellEffectEntry const* effect)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "Spell ScriptStart spellid %u in EffectDummy", m_spellInfo->Id);
-    m_caster->GetMap()->ScriptsStart(DBS_ON_SPELL, m_spellInfo->Id, m_caster, unitTarget);
-}
-
-/**
- * @brief Triggers another spell on the current unit target.
- *
- * @param effect The effect index providing the triggered spell id.
- */
-void Spell::EffectTriggerSpell(SpellEffectEntry const* effect)
-{
-    // only unit case known
-    if (!unitTarget)
-    {
-        if (gameObjTarget || itemTarget)
-        {
-            sLog.outError("Spell::EffectTriggerSpell (Spell: %u): Unsupported non-unit case!", m_spellInfo->Id);
-        }
-        return;
-    }
-
-    uint32 triggered_spell_id = effect->EffectTriggerSpell;
-
-    // special cases
-    switch (triggered_spell_id)
-    {
-        case 18461:                                         // Vanish (not exist)
-        {
-            unitTarget->RemoveSpellsCausingAura(SPELL_AURA_MOD_ROOT);
-            unitTarget->RemoveSpellsCausingAura(SPELL_AURA_MOD_DECREASE_SPEED);
-            unitTarget->RemoveSpellsCausingAura(SPELL_AURA_MOD_STALKED);
-
-            // if this spell is given to NPC it must handle rest by it's own AI
-            if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            uint32 spellId = 1784;
-            // reset cooldown on it if needed
-            if (((Player*)unitTarget)->HasSpellCooldown(spellId))
-            {
-                ((Player*)unitTarget)->RemoveSpellCooldown(spellId);
-            }
-
-            m_caster->CastSpell(unitTarget, spellId, true);
-            return;
-        }
-        case 29284:                                         // Brittle Armor - (need add max stack of 24575 Brittle Armor)
-            m_caster->CastSpell(unitTarget, 24575, true, m_CastItem, NULL, m_originalCasterGUID);
-            return;
-        case 29286:                                         // Mercurial Shield - (need add max stack of 26464 Mercurial Shield)
-            m_caster->CastSpell(unitTarget, 26464, true, m_CastItem, NULL, m_originalCasterGUID);
-            return;
-        case 31980:                                         // Righteous Defense
-        {
-            m_caster->CastSpell(unitTarget, 31790, true, m_CastItem, NULL, m_originalCasterGUID);
-            return;
-        }
-        case 35729:                                         // Cloak of Shadows
-        {
-            Unit::SpellAuraHolderMap& Auras = unitTarget->GetSpellAuraHolderMap();
-            for (Unit::SpellAuraHolderMap::iterator iter = Auras.begin(); iter != Auras.end(); ++iter)
-            {
-                // Remove all harmful spells on you except positive/passive/physical auras
-                if (!iter->second->IsPositive() &&
-                        !iter->second->IsPassive() &&
-                        !iter->second->IsDeathPersistent() &&
-                        (GetSpellSchoolMask(iter->second->GetSpellProto()) & SPELL_SCHOOL_MASK_NORMAL) == 0)
-                {
-                    m_caster->RemoveAurasDueToSpell(iter->second->GetSpellProto()->Id);
-                    iter = Auras.begin();
-                }
-            }
-            return;
-        }
-        case 41967:                                         // Priest Shadowfiend (34433) need apply mana gain trigger aura on pet
-        {
-            if (Unit* pet = unitTarget->GetPet())
-            {
-                pet->CastSpell(pet, 28305, true);
-            }
-            return;
-        }
-        case 58832:                                         // Mirror Image
-        {
-            // Glyph of Mirror Image
-            if (m_caster->HasAura(63093))
-            {
-                m_caster->CastSpell(m_caster, 65047, true, m_CastItem, NULL, m_originalCasterGUID);
-            }
-            break;
-        }
-    }
-
-    // normal case
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(triggered_spell_id);
-    if (!spellInfo)
-    {
-        sLog.outError("EffectTriggerSpell of spell %u: triggering unknown spell id %i", m_spellInfo->Id, triggered_spell_id);
-        return;
-    }
-
-    // select formal caster for triggered spell
-    Unit* caster = m_caster;
-
-    // some triggered spells require specific equipment
-    if (spellInfo->GetEquippedItemClass() >=0 && m_caster->GetTypeId()==TYPEID_PLAYER)
-    {
-        // main hand weapon required
-        if (spellInfo->HasAttribute(SPELL_ATTR_EX3_MAIN_HAND))
-        {
-            Item* item = ((Player*)m_caster)->GetWeaponForAttack(BASE_ATTACK, true, false);
-
-            // skip spell if no weapon in slot or broken
-            if (!item)
-            {
-                return;
-            }
-
-            // skip spell if weapon not fit to triggered spell
-            if (!item->IsFitToSpellRequirements(spellInfo))
-            {
-                return;
-            }
-        }
-
-        // offhand hand weapon required
-        if (spellInfo->HasAttribute(SPELL_ATTR_EX3_REQ_OFFHAND))
-        {
-            Item* item = ((Player*)m_caster)->GetWeaponForAttack(OFF_ATTACK, true, false);
-
-            // skip spell if no weapon in slot or broken
-            if (!item)
-            {
-                return;
-            }
-
-            // skip spell if weapon not fit to triggered spell
-            if (!item->IsFitToSpellRequirements(spellInfo))
-            {
-                return;
-            }
-        }
-    }
-    else
-    {
-        // Note: not exist spells with weapon req. and IsSpellHaveCasterSourceTargets == true
-        // so this just for speedup places in else
-        caster = IsSpellWithCasterSourceTargetsOnly(spellInfo) ? unitTarget : m_caster;
-    }
-
-    caster->CastSpell(unitTarget, spellInfo, true, m_CastItem, NULL, m_originalCasterGUID, m_spellInfo);
-}
-
-/**
- * @brief Triggers a missile spell at the stored destination coordinates.
- *
- * @param effect The effect index providing the triggered spell id.
- */
-void Spell::EffectTriggerMissileSpell(SpellEffectEntry const* effect)
-{
-    uint32 triggered_spell_id = effect->EffectTriggerSpell;
-
-    // normal case
-    SpellEntry const* spellInfo = sSpellStore.LookupEntry(triggered_spell_id);
-
-    if (!spellInfo)
-    {
-        if (unitTarget)
-        {
-            DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "Spell ScriptStart spellid %u in EffectTriggerMissileSpell", m_spellInfo->Id);
-            m_caster->GetMap()->ScriptsStart(DBS_ON_SPELL, m_spellInfo->Id, m_caster, unitTarget);
-        }
-        else
-        {
-            sLog.outError("EffectTriggerMissileSpell of spell %u (eff: %u): triggering unknown spell id %u",
-                      m_spellInfo->Id, effect->EffectIndex, triggered_spell_id);
-        }
-        return;
-    }
-
-    if (m_CastItem)
-    {
-        DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "WORLD: cast Item spellId - %i", spellInfo->Id);
-    }
-
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_caster->CastSpell(m_targets.m_destX, m_targets.m_destY, m_targets.m_destZ, spellInfo, true, m_CastItem, NULL, m_originalCasterGUID, m_spellInfo);
-    }
-    else if (unitTarget)
-    {
-        m_caster->CastSpell(unitTarget, spellInfo, true, m_CastItem, NULL, m_originalCasterGUID, m_spellInfo);
-    }
-}
-
-void Spell::EffectJump(SpellEffectEntry const* effect)
-{
-    if (m_caster->IsTaxiFlying())
-    {
-        return;
-    }
-
-    // Init dest coordinates
-    float x, y, z, o;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-
-        if (effect->EffectImplicitTargetA == TARGET_BEHIND_VICTIM)
-        {
-            // explicit cast data from client or server-side cast
-            // some spell at client send caster
-            Unit* pTarget = NULL;
-            if (m_targets.getUnitTarget() && m_targets.getUnitTarget() != m_caster)
-            {
-                pTarget = m_targets.getUnitTarget();
-            }
-            else if (unitTarget->getVictim())
-            {
-                pTarget = m_caster->getVictim();
-            }
-            else if (m_caster->GetTypeId() == TYPEID_PLAYER)
-            {
-                pTarget = m_caster->GetMap()->GetUnit(((Player*)m_caster)->GetSelectionGuid());
-            }
-
-            o = pTarget ? pTarget->GetOrientation() : m_caster->GetOrientation();
-        }
-        else
-        {
-            o = m_caster->GetOrientation();
-        }
-    }
-    else if (unitTarget)
-    {
-        unitTarget->GetContactPoint(m_caster, x, y, z, CONTACT_DISTANCE);
-        o = m_caster->GetOrientation();
-    }
-    else if (gameObjTarget)
-    {
-        gameObjTarget->GetContactPoint(m_caster, x, y, z, CONTACT_DISTANCE);
-        o = m_caster->GetOrientation();
-    }
-    else
-    {
-        sLog.outError("Spell::EffectJump - unsupported target mode for spell ID %u", m_spellInfo->Id);
-        return;
-    }
-
-    // Try to normalize Z coord because GetContactPoint do nothing with Z axis
-    m_caster->UpdateAllowedPositionZ(x, y, z);
-
-    float speed = m_spellInfo->speed ? m_spellInfo->speed : 27.0f;
-    m_caster->GetMotionMaster()->MoveDestination(x, y, z, o, speed, 2.5f);
-}
-
-void Spell::EffectTeleportUnits(SpellEffectEntry const* effect)   // TODO - Use target settings for this effect!
-{
-    if (!unitTarget || unitTarget->IsTaxiFlying())
-    {
-        return;
-    }
-
-        switch (m_spellInfo->Id)
-        {
-            case 48129:                                 // Scroll of Recall
-            case 60320:                                 // Scroll of Recall II
-            case 60321:                                 // Scroll of Recall III
-            {
-                uint32 failAtLevel = 0;
-                switch (m_spellInfo->Id)
-                {
-                    case 48129: failAtLevel = 40; break;
-                    case 60320: failAtLevel = 70; break;
-                    case 60321: failAtLevel = 80; break;
-                }
-
-                if (unitTarget->getLevel() > failAtLevel && unitTarget->GetTypeId() == TYPEID_PLAYER)
-                {
-                    unitTarget->CastSpell(unitTarget, 60444, true);
-                    // TODO: Unclear use of probably related spell 60322
-                    uint32 spellId = (((Player*)unitTarget)->GetTeam() == ALLIANCE ? 60323 : 60328) + urand(0, 7);
-                    unitTarget->CastSpell(unitTarget, spellId, true);
-                    return;
-                }
-                break;
-            }
-        }
-
-    // Target dependend on TargetB, if there is none provided, decide dependend on A
-    uint32 targetType = effect->EffectImplicitTargetB;
-    if (!targetType)
-    {
-        targetType = effect->EffectImplicitTargetA;
-    }
-
-    switch (targetType)
-    {
-        case TARGET_INNKEEPER_COORDINATES:
-        {
-            // Only players can teleport to innkeeper
-            if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-            {
-                return;
-            }
-
-            ((Player*)unitTarget)->TeleportToHomebind(unitTarget == m_caster ? TELE_TO_SPELL : 0);
-            return;
-        }
-        case TARGET_AREAEFFECT_INSTANT:                     // in all cases first TARGET_TABLE_X_Y_Z_COORDINATES
-        case TARGET_TABLE_X_Y_Z_COORDINATES:
-        {
-            SpellTargetPosition const* st = sSpellMgr.GetSpellTargetPosition(m_spellInfo->Id);
-            if (!st)
-            {
-                sLog.outError("Spell::EffectTeleportUnits - unknown Teleport coordinates for spell ID %u", m_spellInfo->Id);
-                return;
-            }
-
-            if (st->target_mapId == unitTarget->GetMapId())
-            {
-                unitTarget->NearTeleportTo(st->target_X, st->target_Y, st->target_Z, st->target_Orientation, unitTarget == m_caster);
-            }
-            else if (unitTarget->GetTypeId() == TYPEID_PLAYER)
-            {
-                ((Player*)unitTarget)->TeleportTo(st->target_mapId, st->target_X, st->target_Y, st->target_Z, st->target_Orientation, unitTarget == m_caster ? TELE_TO_SPELL : 0);
-            }
-            break;
-        }
-        case TARGET_EFFECT_SELECT:
-        {
-            // m_destN filled, but sometimes for wrong dest and does not have TARGET_FLAG_DEST_LOCATION
-
-            float x = unitTarget->GetPositionX();
-            float y = unitTarget->GetPositionY();
-            float z = unitTarget->GetPositionZ();
-            float orientation = m_caster->GetOrientation();
-
-            m_caster->NearTeleportTo(x, y, z, orientation, unitTarget == m_caster);
-            return;
-        }
-        case TARGET_BEHIND_VICTIM:
-        {
-            Unit* pTarget = NULL;
-
-            // explicit cast data from client or server-side cast
-            // some spell at client send caster
-            if (m_targets.getUnitTarget() && m_targets.getUnitTarget() != unitTarget)
-            {
-                pTarget = m_targets.getUnitTarget();
-            }
-            else if (unitTarget->getVictim())
-            {
-                pTarget = unitTarget->getVictim();
-            }
-            else if (unitTarget->GetTypeId() == TYPEID_PLAYER)
-            {
-                pTarget = unitTarget->GetMap()->GetUnit(((Player*)unitTarget)->GetSelectionGuid());
-            }
-
-            // Init dest coordinates
-            float x = m_targets.m_destX;
-            float y = m_targets.m_destY;
-            float z = m_targets.m_destZ;
-            float orientation = pTarget ? pTarget->GetOrientation() : unitTarget->GetOrientation();
-            unitTarget->NearTeleportTo(x, y, z, orientation, unitTarget == m_caster);
-            return;
-        }
-        default:
-        {
-            // If not exist data for dest location - return
-            if (!(m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION))
-            {
-                sLog.outError( "Spell::EffectTeleportUnits - unknown EffectImplicitTargetB[%u] = %u for spell ID %u", effect->EffectIndex, effect->EffectImplicitTargetB, m_spellInfo->Id );
-                return;
-            }
-            // Init dest coordinates
-            float x = m_targets.m_destX;
-            float y = m_targets.m_destY;
-            float z = m_targets.m_destZ;
-            float orientation = unitTarget->GetOrientation();
-            // Teleport
-            unitTarget->NearTeleportTo(x, y, z, orientation, unitTarget == m_caster);
-            return;
-        }
-    }
-
-    // post effects for TARGET_TABLE_X_Y_Z_COORDINATES
-    switch (m_spellInfo->Id)
-    {
-            // Dimensional Ripper - Everlook
-        case 23442:
-        {
-            int32 r = irand(0, 119);
-            if (r >= 70)                                    // 7/12 success
-            {
-                if (r < 100)                                // 4/12 evil twin
-                {
-                    m_caster->CastSpell(m_caster, 23445, true);
-                }
-                else                                        // 1/12 fire
-                {
-                    m_caster->CastSpell(m_caster, 23449, true);
-                }
-            }
-            return;
-        }
-        // Ultrasafe Transporter: Toshley's Station
-        case 36941:
-        {
-            if (roll_chance_i(50))                          // 50% success
-            {
-                int32 rand_eff = urand(1, 7);
-                switch (rand_eff)
-                {
-                    case 1:
-                        // soul split - evil
-                        m_caster->CastSpell(m_caster, 36900, true);
-                        break;
-                    case 2:
-                        // soul split - good
-                        m_caster->CastSpell(m_caster, 36901, true);
-                        break;
-                    case 3:
-                        // Increase the size
-                        m_caster->CastSpell(m_caster, 36895, true);
-                        break;
-                    case 4:
-                        // Decrease the size
-                        m_caster->CastSpell(m_caster, 36893, true);
-                        break;
-                    case 5:
-                        // Transform
-                    {
-                        if (((Player*)m_caster)->GetTeam() == ALLIANCE)
-                        {
-                            m_caster->CastSpell(m_caster, 36897, true);
-                        }
-                        else
-                        {
-                            m_caster->CastSpell(m_caster, 36899, true);
-                        }
-                        break;
-                    }
-                    case 6:
-                        // chicken
-                        m_caster->CastSpell(m_caster, 36940, true);
-                        break;
-                    case 7:
-                        // evil twin
-                        m_caster->CastSpell(m_caster, 23445, true);
-                        break;
-                }
-            }
-            return;
-        }
-        // Dimensional Ripper - Area 52
-        case 36890:
-        {
-            if (roll_chance_i(50))                          // 50% success
-            {
-                int32 rand_eff = urand(1, 4);
-                switch (rand_eff)
-                {
-                    case 1:
-                        // soul split - evil
-                        m_caster->CastSpell(m_caster, 36900, true);
-                        break;
-                    case 2:
-                        // soul split - good
-                        m_caster->CastSpell(m_caster, 36901, true);
-                        break;
-                    case 3:
-                        // Increase the size
-                        m_caster->CastSpell(m_caster, 36895, true);
-                        break;
-                    case 4:
-                        // Transform
-                    {
-                        if (((Player*)m_caster)->GetTeam() == ALLIANCE)
-                        {
-                            m_caster->CastSpell(m_caster, 36897, true);
-                        }
-                        else
-                        {
-                            m_caster->CastSpell(m_caster, 36899, true);
-                        }
-                        break;
-                    }
-                }
-            }
-            return;
-        }
-    }
-}
-
-/**
  * @brief Creates and attaches an aura effect to the current unit target.
  *
  * @param effect The aura effect index.
